requirements.txt
python-telegram-bot>=12.8
emoji>=0.6.0
requests>=2.24.0
pyyaml>=5.4.1
